+++ src/mare/Main.cpp
@@ -196,7 +196,7 @@ int main(int argc, char* argv[])
       else
       {
         const char* arg = argv[i] + 2;
-        unsigned int arglen = strlen(arg);
+        size_t arglen = strlen(arg);
         const char* argarg = strchr(arg, '=');
         if(argarg)
           arglen = argarg - arg;
